backend/requirements.txt
flask
requests
BeautifulSoup4
pillow
pytesseract
pydantic